LayoutTests/fast/webgpu/primitive-index-expected.txt

Tests the primitive_index builtin from the WGSL primitive_index extension. Renders 4 non-overlapping triangles in a single draw call and uses @builtin(primitive_index) in the fragment shader to write each primitive's index into a storage buffer indexed by pixel position. Verifies the index seen by the fragment shader matches the draw order.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS Adapter advertises 'primitive-index' feature
Per-pixel primitive_index+1: 1,2,3,4
PASS primitiveIndexBuf[0] is 1
PASS primitiveIndexBuf[1] is 2
PASS primitiveIndexBuf[2] is 3
PASS primitiveIndexBuf[3] is 4
PASS successfullyParsed is true

TEST COMPLETE
